+++ src/data/codec/deser/mod.rs
@@ -1,23 +1,19 @@
 // devela::data::codec::deser
 //
-#![doc = crate::_tags!(wip)]
-//! Structured serialization/deserialization.
+#![doc = crate::_DOC_DATA_CODEC!()] // public
+#![doc = crate::_doc!(modules: crate::data; codec: crypto, hash)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: hash)]
+#![doc = crate::_QUO_DATA_CODEC!()]
 //
+// > how structured values are assembled/disassembled
 
-// mod byte; // ByteSink, ByteSource WIP
-// mod macros; // WIP IDEA
-// mod rle; // WIP
-// mod traits; // Deserialize, Serialize… WIP
-// mod utils; // WIP
+// mod traits; // WIP
 
 crate::structural_mods! { // _mods
     _mods {
         // pub use super::{
-        //     byte::*,
-        //     // macros::*,
-        //     // rle::*,
         //     traits::*,
-        //     // utils::_all::*,
         // };
     }
 }