+++ src/data/codec/frame/mod.rs
@@ -1,0 +1,16 @@
+// devela::data::codec::frame
+//
+#![doc = crate::_DOC_DATA_CODEC_FRAME!()] // public
+#![doc = crate::_doc!(modules: crate::data::codec; frame)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: frame)]
+//
+// > Where a byte unit begins and ends
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     __::_all::*,
+        // };
+    }
+}