+++ src/data/codec/schema/mod.rs
@@ -1,0 +1,16 @@
+// devela::data::codec::schema
+//
+#![doc = crate::_DOC_DATA_CODEC_SCHEMA!()] // public
+#![doc = crate::_doc!(modules: crate::data::codec; schema)]
+#![doc = crate::_doc!(flat:"data")]
+#![doc = crate::_doc!(extends: schema)]
+//
+// > What fields/kinds exist inside a value or record
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     __::_all::*,
+        // };
+    }
+}